finish lesson plan and slides for lesson 12 on apis; finish exercises for lesson 12
