+++ 12-apis/exercise/exercise3/package.json
@@ -8,10 +8,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.14.0",
-    "express-handlebars": "^3.0.0",
-    "handlebars-form-helpers": "^0.1.3",
-    "hbs": "^4.0.1"
+    "express": "^4.14.0"
   },
   "devDependencies": {
     "nodemon": "^1.10.2"
