--- 12-apis/exercise/exercise4/css/style.css
@@ -1,27 +1,0 @@
-* {
-  font-family: 'helvetica neue';
-}
-
-body {
-  height: 100%;
-  width: 100%;
-  padding: 5em 0;
-}
-
-form {
-  text-align: center;
-}
-
-.content {
-  width: 60%;
-  max-width: 700px;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-h1 {
-  text-align: center;
-}
