--- 12-apis/exercise/exercise4/views/home.handlebars
@@ -1,7 +1,0 @@
-<h1>Pokedex</h1>
-
-<h3>All Pokemon:</h3>
-
-{{#each pokemon}}
-  {{id}} - <a href="/pokemon/{{id}}">{{name}}</a> <br>
-{{/each}}
