--- 12-apis/exercise/exercise4/views/layouts/main.handlebars
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>Our First Application</title>
-</head>
-<body>
-
-    {{{body}}}
-
-</body>
-</html>
