--- 12-apis/exercise/exercise4/views/pokemon.handlebars
@@ -1,36 +1,0 @@
-<h1>{{name}}</h1>
-
-<table>
-  <tr>
-    <th>
-      Attack:
-    </th>
-    <td>
-      {{attack}}
-    </td>
-  </tr>
-  <tr>
-    <th>
-      Defense:
-    </th>
-    <td>
-      {{defense}}
-    </td>
-  </tr>
-  <tr>
-    <th>
-      Evolve Level:
-    </th>
-    <td>
-      {{evolveLevel}}
-    </td>
-  </tr>
-  <tr>
-    <th>
-      type:
-    </th>
-    <td>
-      {{type}}
-    </td>
-  </tr>
-</table>
