--- 12-apis/exercise/exercise4/views/search.handlebars
@@ -1,9 +1,0 @@
-<div class="content">
-  <h1>Search the Pokedex</h1>
-  <form id="searchForm">
-    <input type="text" name="name" placeholder="Pokemon Name" />
-    <input type="text" name="moves" placeholder="Special Moves"/>
-    <input type="text" name="probability" placeholder="Probability"/>
-    <input type="submit" value="Submit" onclick="pokemonSearch();" />
-  </form>
-</div>
