+++ 12-apis/slides.md
@@ -1,22 +1,21 @@
 
-
 <img src="https://ga-core.s3.amazonaws.com/production/uploads/program/default_image/5225/JS-logo-official.png" style="max-width: 100px; border: none; box-shadow: none" />
 ## Class 12: APIs
 
 ---
 ### Agenda
-| Timing | Topic                                    |
-| ------ | ---------------------------------------- |
-| 10 min | Review HTTP                              |
-| 30 min | HTTP Warmup Exercise:                    |
-| 15 min | AJAX                                     |
-| 15 min | AJAX Exercise                            |
-| 10 min | Break                                    |
-| 15 min | Introducing APIs                         |
-| 15 min | OMDB                    |
-| 25 min | Code along Weatherify                    |
-| 30 min | Final Exercise                           |
-
+| Timing | Topic                     |
+| ------ | ------------------------- |
+| 5  min | Check In                  |
+| 30 min | Warmup                    |
+| 20 min | Introduction to AJAX      |
+| 10 min | Break                     |
+| 20 min | Working with AJAX         |
+| 30 min | Working with APIs         |
+| 10 min | Break                     |
+| 20 min | Working with APIs         |
+| 30 min | Building an API cont'd    |
+| 5  min | Closing Questions & Exit  |
 
 ---
 ## Looking Ahead
@@ -25,52 +24,68 @@
 | ---------- | ------------------------ |
 |     07     |   Intro to the DOM       |
 |     08     | Events                   |
-|     09     | Templating & Build Tools |
+|     09     | Templating               |
 |     10     | Express                  |
-|     11     | Databases & CRUD       |
-|   **11**   | **APIs**                   |
+|     11     | Databases & CRUD         |
+|   **11**   | **APIs**                 |
 |     12     | Application Architecture |
 |     13     | Lab #2                   |
 
---
-### HTTP recap
-- Client and Server relationship
-- HTTP is their love language
-
---
+---
 ### Warmup Exercise
 
 --
-### The History of Ajax
-- Ajax is open source and thus accessible
-- The term AJAX was not used until 2005, but development had been underway since around 1996
-- It can’t be credited to just one organization or individual.
+## AJAX
 
-### Getting to know AJAX
+--
+### Introduction to Ajax
+- Way for the client to communicate with the server
+- Doesn't require a page refresh
+
+--
+### The History of Ajax
 - Ajax is not a programming language or a tool, but a concept.
-- Ajax is a client-side script that communicates to and from a server/database without the need for a postback or a complete page refresh
+- The term AJAX was not used until 2005, but development had been underway since around 1996
 - More narrowly, it refers to the use of XmlHttpRequest objects to interact with a web server dynamically via JavaScript.
 
+--
 ### The 4 Benefits of AJAX
 1. Callbacks
-- Ajax is used to perform a callback, making a quick round trip to and from the server to retrieve and/or save data without posing the entire page back to the server
-- What benefits could this have?
 2. Asynchronous Calls
-- Ajax allows you to make Asynchronous calls to a web server
-- The browser can avoid waiting for all data to arrive before allowing the user to act once more
 3. User-Friendly
-- Ajax enabled applications will always be more responsive, faster and more User-Friendly
 4. Increase Speed
 
+--
+### 1. Callbacks
+- Ajax is used to perform a callback, making a quick round trip to and from the server to retrieve and/or save data without posting the entire page back to the server
+- What benefits could this have?
+
+--
+### 2. Asynchronous Calls
+- Ajax allows you to make Asynchronous calls to a web server
+- The browser can avoid waiting for all data to arrive before allowing the user to act
+
+--
+### 3. User-Friendly
+- Ajax enabled applications will always be more responsive, feel faster and thus provide a better user experience
+
+--
+### 4. Increase Speed
+- rather than require the user to navigate to entirely different HTML document, we can just get the data we need using AJAX and update the current document
+
+--
 ### Where can we use Ajax?
 - Ajax should be used anywhere in a web application where small amounts of information could be saved or retrieved from the server without posting back the entire pages
 - Can we think of any examples?
+
 --
 
 ### Lets explore the technical aspects of Ajax
 Exercise 1
+
 ---
 ### Introducing APIs
+
 --
 
 ### A Brief History of APIs
